+++ happytransformer/adaptors/__init__.py
@@ -1,1 +1,16 @@
-from .adaptor import get_adaptor
+from .adaptor import Adaptor
+from .albert import AlbertAdaptor
+from .berts import BertAdaptor,DistilBertAdaptor,RobertaAdaptor
+
+ADAPTORS = {
+    'BERT':BertAdaptor(),
+    'DISTILBERT':DistilBertAdaptor(),
+    'ROBERTA':RobertaAdaptor(),
+    'ALBERT':AlbertAdaptor()
+}
+
+def get_adaptor(model_type:str)->Adaptor:
+    if model_type in ADAPTORS:
+        return ADAPTORS[model_type]
+    else:
+        raise ValueError(f'Model type <{model_type}> not currently supported')
